Transpose @romainfd fonctions into the trigger step and utils directly
